refactoring: BindableEditor
